PCI/P2PDMA: Split out the information about the providing device from pgmap
The main thing that P2PDMA stores in the pgmap is the providing struct
device and the computed bus_offset for non-host transactions.

Introduce a new stand alone structure 'p2pdma_provider' that stores this
information and embed it in the struct pci_p2pdma_pagemap to replace the
original information.

After the next patch all the calculations related to DMA mapping P2PDMA
memory will be done only against p2pdma_provider and pgmap simply turns
into a way to locate the p2pdma_provider pointer.

The intention is to make the P2PDMA system usable without a pgmap or
struct page.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>