dma: Document what min_align_mask is for
The only places that set it currently use 0xfff, and the only
implementation is in swiotlb that enforces that the DMA address
has those bits matching the physical bits.

It looks like the purpose is to allow the low bits of the address to
be the same after mapping, eg the offset within a HW structure is
constant.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>