lib/rlist: Introduce range list
range list is a general data structure the holds a logical linear list of
start/length/extra ranges. It is a tool to solve a similar problem as a
'struct page *' array, bio_vec or scatterlist.

Specifically:
 - The sum of lengths in the list is "the position"

 - Each position from [0,total size) can be assigned to exactly one range
   entry

 - Mergable adjoining ranges can be merged with no loss in information -
   unlike scatterlist the range boundaries are not information

 - The list can only be iterated linearly, no efficient seeking

Each stored range has:
 - 96 bits to describe the 'base' (64 bit base and 32 bit offset)
 - 60 bits for the length
 - 34 bits of 'extra' data that is for the data structure user

The rlist focuses on memory efficiency and flexibility of the underlying
storage. It uses an xarray-like dynamically growable fixed-chunk scheme,
with a variable 8/16/24 byte entry encoding. It offers an opaque API to
the user that hides these implementation details.

The automatic entry sizing means a single data structure can be space
optimal for 32 and 64 bit systems, as it automatically sizes based on what
is stored within it.

The idea is the user will adjust what it is storing to fit into the
smaller entry sizes, following patches use a few strategies:

 - folios are stored as 'struct folio *' pointers to avoid
   compound_head() related cache dirtying

 - full folios can be encoded as a PFN and a length in PFN, this allows
   page lists to use 8 bytes per folio

 - dma_addr_t is relative to the starting IOVA when using a IOMMU,
   thus most entries are 8 bytes.

 - Things that can't be optimized just use the 24 byte entry format

This is likely only an internal data structure, it is quirky enough that
few places would want to use it directly. Keeping this storage code
seperated from the next patches makes it easier to understand and
maintain, and perhaps someday replace completely.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>