iommu: WIP imagine the iommu drivers can use rlist
There is alot of overhead in the iommu scheme when working with lists of
memory - we have to call a function pointer for every contiguous page and
the driver has to chase down ~4-5 levels of page tables to get back to the
next IOPTE spot to stuff.

This could all be a lot better if the driver could iterate and segment
the physaddr_t list.

Further, we can teach things like iova_to_phys to be much faster and
natively return the large pages stored in the iommu instead of querying
each 4K chunk.

Finally, we can imagine an umap that reads the physical pages before
invalidating the IOPTE. This would be a big help for iommufd as it uses
the IOPTEs to store the physical pages and it does a lot of slow work to
read them out before unmapping.

We can imagine the APIs here but implementing them would likely require a
major consolidation in how the drivers process IOPTE to allow generic code
to exist.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>