lib/rlist: Introduce rlist_dma
This is the output side of the DMA API when working with the rlist cpu
iterator. It provides a list of dma_addr_t's that can be programmed into
device hardware, along with a number of useful helpers to aid drivers in
building their HW programming.

It is distinct from the rlist_cpu for type safety. scatterlist has shown
there is significant problems with type confusion when allowing the two
APIs to cross.

The rlist_dma can store the underlying ranges in two ways:

 - A pointer to an rlist cpu iterator - in this case the dma_addr and the
   phys_addr are direct mapped. This allows a common case of the DMA API's
   direct map function to simply point at the CPU iterator and do nothing
   else

 - A 'struct rlist' which will store the dma_addr_t's and opaque data for
   the dma_map_ops implementation natively in rlist entries.

As with the rlist CPU iterator the exposed iterator API hides these
details.

To support HW programming the rlist_dma has iterators that segment
according to several common segmentation rules.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>