dma: Add DMA API support for mapping a rlist_cpu to a rlist_dma
This API is the big purpose of this work.

Allowing the DMA API to efficiently map a data structure that is not
scatterlist is essential to introducing a new data structure anywhere else
in the driver stack. At some point every HW driver needs to DMA map the
memory it is working with.

This API is modeled after the scatterlist map_sg(). It takes in a
rlist_cpu that specifies the CPU view of the memory ranges and emits a
rlist_dma that specifies the dma_addr_t ranges that cover the same memory.

Like the scatterlist APIs this has a mechanism to manage
segmentation. Since rlist/cpu/dma do not store segmentation information in
the storage the visibility of segmentation while mapping is not essential
for the basic functionality. The actual segmentation will occur when the
HW driver iterates over the rlist for final programming.

Since rlist segmentation includes per-map items for page list HW that have
an impact on IOVA assignment the segmentation information is passed as an
argument. The global device parameters are not used in the native rlist
DMA API implementations. This also allows complex HW with different types
of DMA engines to use engine specific segmentation, eg page list vs
scatter list, or even engine specific limitations on scatter lists.

Segmentation is also used for the usual things the SG path uses it for:

 - To precompute the number of segments/etc in the DMA list because
   driver's can benefit from this information
   FIXME: this hasn't been typed in yet

 - Avoid increasing the number of segments compared to the CPU list.
   This can only be done while assigning IOVAs to the mappings, otherwise
   there are corner cases that a bad IOVA assignment will increase
   the segment count.

   This preservation is useful if the total number of segments is being
   managed in the CPU list domain prior to DMA mapping.

   (The initial rlist users I have planned out do not require this)

The parameters assigned to the struct device should continue to be the
worst case segmentation that a driver will need, however drivers can
freely use more relaxed segmentation rules on a per-HW unit basis. This is
because the following patches do not covert every architecture to native
rlist mode so a few architecture do rely on the per-device parameters
being a worst case.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>