cover-letter: RFC Create an alternative to scatterlist in the DMA API
This was kicked off by Matthew with his phyr idea from this thread:

https://lore.kernel.org/linux-mm/YdyKWeU0HTv8m7wD@casper.infradead.org/

Hwoevr, I have become interested in this with some immediacy because of
IOMMUFD and this other discussion with Christoph:

https://lore.kernel.org/kvm/4-v2-472615b3877e+28f7-vfio_dma_buf_jgg@nvidia.com/

Which results in, more or less, we have no way to do P2P DMA operations
without struct page. This becomes complicated when we touch RDMA which
highly relies on scatterlist for its internal implementations, so being
unable to use scatterlist to store only dma_addr_t's means RDMA needs a
complete scatterlist replacement that can.

So - my objective is to enable to DMA API to "DMA map" something that is
not a scatterlist, may or may not contain struct pages, but can still
contain P2P DMA physical addresses. With this tool, transform the RDMA
subystem to use the new DMA API and then go into DMABUF and stop creating
scatterlists without any CPU pages. From that point we could implement
DMABUF in VFIO (as above) and use the DMABUF to feed the MMIO pages into
IOMMUFD to restore the PCI P2P support in VMs withotu creating the
follow_pte security problem that VFIO has.

After going through the thread again, and making some sketches, I've come
up with this suggestion as a path forward, explored very roughly in this
RFC:

1) Create something I've called a 'range list CPU iterator'. This is an
   API that abstractly iterates over CPU physical memory ranges. It
   has useful helpers to iterate over things in 'struct page/folio *',
   physical ranges, copy to/from, and so on

   It has the necessary extra bits beyond the physr sketch to support P2P
   in the DMA API based on what was done for the pgmap based stuff. ie we
   need to know the provider of the non-struct page memory to get the
   struct device to compute the p2p distance and compute the pci_offset.

   The immediate idea is this is an iterator, not a data structure. So it
   can iterate over different kinds of storage. This frees us from having
   to immediatly consolidate all the different storage schemes in the
   kernel and lets that work happen over time.

   I imagine we would want to have this work with struct page * (for GUP)
   and bio_vec (for storage/net) and something else for the "kitchen sink"
   with DMABUF/etc. We may also want to allow it to wrapper scatterlist to
   provide for a more gradual code migration.

   Things are organized so sometime in the future this could collapse down
   into something that is not a multi-storage iterator, but perhaps just
   a single storage type that everyone is happy with.

   In the mean time we can use the API to progress all the other related
   infrastructure.

   Fundamentally this tries to avoid the scatterlist mistake of leaking
   too much of the storage implementation detail to the user.

2) Create a general storage called the "range list". This is intended to
   be a general catch-all like scatterlist is, and it is optimized
   towards page list users, so it is quite good for what RDMA wants.

3) Create a "range list DMA iterator" which is the dma_addr_t version of
   #1. This needs to have all the goodies to break up the ranges into
   things HW would like, such as page lists, or restricted scatter/gather
   records.

   I've been able to draft several optimizations in the DMA mapping path
   that should help offset some of the CPU cost of the more abstracted
   iterators:

       - DMA direct can directly re-use the CPU list with no iteration or
         memory allocation.

       - The IOMMU path can do only one iteration by pre-recording if the
         CPU list was all page aligned when it was created

The following patches go deeper into my thinking, present fairly complete
drafts of what things could look like, and more broadly explores the whole
idea.

At the end of the series we have
 - rlist, rlist_cpu, rlist_dma implementations
 - An rlist implementation for every dma_map_ops
 - Good rlist implementations for DMA direct and dma-iommu.c
 - A pin_user_pages() wrapper
 - RDMA umem converted and compiling with some RDMA drivers
 - Compile tested only :)

It is a huge amount of work, I'd like to get a sense of what people think
before going more deepely into a more final tested implementation. I know
this is not quite what Matthew and Christoph have talked about.

Cc: Christoph Hellwig <hch@lst.de>
Cc: Matthew Wilcox <willy@infradead.org>
Cc: Shakeel Butt <shakeelb@google.com>
Cc: Chaitanya Kulkarni <chaitanyak@nvidia.com>
Cc: Bart Van Assche <bvanassche@acm.org>
Cc: Mike Rapoport <rppt@kernel.org>
Cc: Zhu Yanjun <yanjun.zhu@linux.dev>
Cc: Christian König <christian.koenig@amd.com>
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>