dma: Add DMA direct support for rlist mapping
The DMA direct rlist functions are very similar to the SG ones, they use
the 'dma_map_ops_priv' to encode if each entry is a P2P entry or a direct
mapped one.

A fast past is created if we can predict in advance that the mapping
doesn't require any special things. The fast path allows the rlist_dma to
point directly at the rlist_cpu and avoids memory allocation and
iteration. It relies on being able to predict this from the device
attributes.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>