PCI/P2PDMA: Store the p2pdma_provider structs in an xarray
Generate a small integer than can be used as a handle for the
p2pdma_provider struct in the rlist_cpu since we don't have room for a
full 64 bit pointer.

FIXME: A small ID goes nicely along with the design of the rlist that has
this extra space to store it. An alternative is to have an interval_tree
here and use 1 bit to indicate if a provider search on physical address is
required.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>